MAINT: Renamed module mutual_info to mutual_info_
